third_party/blink/web_tests/external/wpt/html/semantics/forms/the-selectmenu-element/selectmenu-shadow-root-replacement.tentative-expected.txt

This is a testharness.js-based test.
FAIL Test that the UA provided shadow root can be replaced assert_equals: Ensure that controller code has been provided expected true but got false
PASS Test that only the UA provided shadow root can be replaced
Harness: the test ran to completion.
